backend/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
amqp==5.1.1
anyio==4.0.0
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
backoff==2.2.1
billiard==4.1.0
black==23.3.0
boltons==19.3.0
cachetools==5.3.3
celery==5.3.0
certifi==2022.12.7
cffi==1.15.1
cfgv==3.3.1
chardet==5.2.0
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cron-descriptor==1.4.0
cryptography==40.0.1
dadata==21.10.1
DataProperty==0.55.0
defusedxml==0.7.1
dict2xml==1.7.5
distlib==0.3.8
Django==4.2.2
django-allauth==0.54.0
django-appconf==1.0.5
django-celery-beat==2.5.0
django-ckeditor==6.7.1
django-colorfield==0.8.0
django-configurations==2.4.1
django-cors-headers==3.14.0
django-filter==23.1
django-jazzmin==2.6.0
django-js-asset==2.1.0
django-mptt==0.14.0
django-pgbulk==1.3.0
django-pgtrigger==4.11.1
django-pytest==0.2.0
django-smsru==1.0.4
django-solo==2.0.0
django-timezone-field==5.1
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.0
docopt==0.6.2
drf-yasg==1.21.5
easy-thumbnails==2.8.5
entrypoints==0.4
exceptiongroup==1.1.1
filelock==3.13.4
flake8==4.0.1
flakeheaven==3.2.1
frozenlist==1.3.3
fs-cdek-api==1.4.1
fs-moysklad-api==0.5.1
gitdb==4.0.10
GitPython==3.1.31
h11==0.14.0
httpcore==1.0.1
httpx==0.25.1
icalendar==5.0.4
identify==2.5.22
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
isort==5.12.0
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.17.3
kombu==5.3.0
lxml==5.2.0
Markdown==3.4.3
MarkupSafe==2.1.2
mbstrdecoder==1.1.2
mccabe==0.6.1
multidict==6.0.4
mypy==1.2.0
mypy-extensions==1.0.0
nodeenv==1.7.0
oauthlib==3.2.2
packaging==24.0
pathspec==0.11.1
pathvalidate==2.5.2
Pillow==9.5.0
piptools-sync==1.0.2
platformdirs==4.2.0
pluggy==1.5.0
pre-commit==3.2.2
pre-commit-hooks==4.4.0
prompt-toolkit==3.0.38
psycopg2-binary==2.9.6
pycodestyle==2.8.0
pycparser==2.21
pydantic==1.10.7
pyflakes==2.4.0
Pygments==2.15.0
PyJWT==1.7.1
pyproject-api==1.6.1
pyrsistent==0.19.3
pytablewriter==0.64.2
pytest==7.3.0
pytest-django==4.8.0
pytest-github-report==0.0.1
python-crontab==2.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0
redis==4.5.5
requests==2.28.2
requests-oauthlib==1.3.1
restdoctor==0.0.61
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
semver==2.13.0
six==1.16.0
smmap==5.0.0
smsru==0.1.0
sniffio==1.3.0
sqlparse==0.4.3
structlog==23.2.0
swagger-spec-validator==3.0.3
tabledata==1.3.1
tcolorpy==0.1.2
toml==0.10.2
tomli==2.0.1
tox==4.14.2
tqdm==4.65.0
typepy==1.3.0
typing_extensions==4.5.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.15
vine==5.0.0
virtualenv==20.25.3
wcwidth==0.2.6
XlsxWriter==3.0.9
yarg==0.1.9
yarl==1.8.2
